% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\encoding{UTF-8}
\name{simulate}
\alias{simulate}
\title{Simulation scenario used in the paper}
\usage{
simulate(n, d = 2, m = 2, rho = 0.5, normalise = FALSE)
}
\arguments{
\item{n}{number of rows}

\item{d}{number of columns}

\item{m}{number of matrices}

\item{rho}{correlation parameter}

\item{normalise}{whether to ensure a V=0 solution exists}
}
\description{
simulate multiple matrices
}
\references{
Cornilly, D., Puccetti, G., Rüschendorf, L., & Vanduffel, S. (2021). 
On a synchronization problem with multiple instances.
}
\author{
Dries Cornilly
}
\concept{simulate}
