\name{SPMI-package}
\alias{SPMI-package}
\alias{SPMI}
\docType{package}
\title{
\packageTitle{SPMI}
}
\description{
\packageDescription{SPMI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SPMI}
\packageIndices{SPMI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SPMI}

Maintainer: \packageMaintainer{SPMI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
