% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra.R
\encoding{UTF-8}
\name{rra}
\alias{rra}
\title{Rearrange multiple matrices}
\usage{
rra(
  x,
  algo = "swap",
  maxiter = 1e+05,
  maxnoimprove = 1000,
  eps = 1e-06,
  p_block = 0.5,
  n_preprocess = 3,
  p_vec = NULL
)
}
\description{
wrapper function for rearrangement of multiple matrices
}
\details{
TODO
}
\references{
Cornilly, D., Puccetti, G., Rüschendorf, L., & Vanduffel, S. (2021). 
On a synchronization problem with multiple instances.
}
\author{
Dries Cornilly
}
\concept{rra}
