% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra.R
\encoding{UTF-8}
\name{rra}
\alias{rra}
\title{Rearrange multiple matrices}
\usage{
rra()
}
\description{
wrapper function for rearrangement of multiple matrices
}
\details{
TODO
}
\references{
Cornilly, D., Puccetti, G., Rüschendorf, L., & Vanduffel, S. (2021). 
On a synchronization problem with multiple instances.
}
\author{
Dries Cornilly
}
\concept{rra}
